Agent/多模态创意生成/requirements.txt
langchain
langchain-core
langchain-openai
openai
python-dotenv
requests
